Consumer/requirements.txt
pymongo
kafka_python
